+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to the code will be documented in this file.
 
 ## 1.1.0
 
-Other Changes
+Features
 
+- Added Save current color as a preferred color
 - When a setting is changed, Peacock should update the colors appropriately based on the most recently used color during the active VS Code instance's session
 
 ## 1.0.0